Adds a script with cronjob to the dev2 server
The script grabs data from the 360 DKAN, fetches csv files, and then
runs a script over those files and outputs the results as txt. It then
transforms the text to html.

This salt set up, creates a user, grabs the code from GitHub, sets
up the crobjob - with logging, and sets up apache